+++ CHANGELOG.md
@@ -1,9 +1,12 @@
-## [2.3.3] - 4/21/26
+## [2.3.4] - 4/21/26
 ### Contracts
 - Updated all 4 Contract templates to the new '2026' format
     - Wording, spacing, colors, & category names corrected
 - Replaced 'Auxiliary Pump I/II' with 'Additional Pump I/II'
     - Blower no longer maps to the Auxiliary Pump line item
+### Warranty Breakdown
+- Editing the Warranty Breakdown fields no longer delete text while typing
+    - This was caused from auto-saving, a 'Save Changes' button has been introduced
 -----
 ## [2.3.2] - 4/20/26
 ### Hotfix